Signout api and admin dashboard

+++ admin-app/src/actions/auth.actions.js
@@ -43,7 +43,7 @@ export const signup = (user) => {
 
     if (res.status === 200) {
       const { token, user } = res.data;
-     
+
       dispatch({
         type: authConstants.LOGIN_SUCCESS,
         payload: {
@@ -85,9 +85,19 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
-    dispatch({
-      type: authConstants.LOGOUT_REQUEST,
-    });
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post(`/admin/signout`);
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
